+++ Changelog.md
@@ -1,1 +1,7 @@
+# v0.1.1, 2021-03
+
+Rust:
+- Configurable field order (`zki example --field-order=101`)
+- Example with incorrect witnes (`zki example --incorrect`)
+
 # v0.1.0, 2020-09